requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
pdfplumber==0.10.2
nltk==3.8.1
reportlab==4.0.9
python-dotenv==1.0.0
groq==0.18.0
